check rpc server when proxy
