+++ src/utils/mod.rs
@@ -1,0 +1,2 @@
+/// ssh utils
+pub mod ssh;
